Added logout functionality and moved tests from App.tsx to Dashboard.test.tsx
